mock: add mock constructor to BlockHeader behind testing feature
