+++ objects/Cargo.toml
@@ -23,7 +23,7 @@ default = ["std"]
 concurrent = ["std"]
 serde = ["dep:serde", "miden-crypto/serde"]
 std = ["assembly/std", "miden-crypto/std", "miden-verifier/std", "vm-core/std", "vm-processor/std"]
-testing = []
+testing = ["dep:winter-rand-utils"]
 
 [dependencies]
 assembly = { workspace = true }
@@ -33,6 +33,7 @@ miden-verifier = { workspace = true }
 serde = { version = "1.0", optional = true, default-features = false, features = ["derive"] }
 vm-core = { workspace = true }
 vm-processor = { workspace = true }
+winter-rand-utils = { version = "0.8", optional = true }
 
 [dev-dependencies]
 criterion = { version = "0.5", default-features = false, features = ["html_reports"] }
